08_conv_AI/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Markdown==3.5.1
pandas==2.1.2
Pillow==10.0.1
plotly==5.18.0
streamlit==1.34.0
yfinance

google-generativeai

google-cloud-aiplatform #this is required
google-cloud-logging  #this is required
google-cloud-storage
fsspec
gcsfs
pyarrow
fastparquet
cachetools
stockstats